Refactor uploader tests
