+++ uploader/test/test_file_upload.py
@@ -8,7 +8,7 @@ from uploader.constants import MANDATORY_SHEETS
 from uploader.models import CofkCollectWork, CofkCollectAuthorOfWork, \
     CofkCollectAddresseeOfWork, CofkCollectOriginOfWork, CofkCollectDestinationOfWork, CofkCollectManifestation
 from uploader.spreadsheet import CofkUploadExcelFile
-from uploader.test.test_utils import UploadIncludedTestCase
+from uploader.test.test_utils import UploadIncludedTestCase, spreadsheet_data
 from uploader.validation import CofkExcelFileError
 
 log = logging.getLogger(__name__)
@@ -78,19 +78,7 @@ class TestFileUpload(UploadIncludedTestCase):
         """
         This test should run successfully as all required data is present and valid.
         """
-        data = {'Work': [[1,"test","J",1660,1,1,1660,1,2,1,1,1,1,"test","newton",15257,"test",1,1,"test","Wren",
-                           22859,"test",1,1,"test","Burford",400285,"test",1,1,"Carisbrooke",782,"test",1,1,"test",
-                           "test","fra;eng",'','','','','','',"test","test","test","Baskerville",885,"test",
-                           "test","EMLO","http://emlo.bodleian.ox.ac.uk/","Early Modern Letters Online test"]],
-                'People': [["Baskerville", 885],
-                           ["newton", 15257],
-                           ["Wren", 22859]],
-                'Places': [['Burford', 400285],
-                           ['Carisbrooke', 782]],
-                'Manifestation': [[1, 1, "ALS", 1, "Bodleian", "test", "test",'','','','',''],
-                                  [2, 1,'','','','','', "P", "test", "test",'','']],
-                'Repositories': [['Bodleian', 1]]}
-        filename = self.create_excel_file(data)
+        filename = self.create_excel_file(spreadsheet_data)
 
         cuef = CofkUploadExcelFile(self.new_upload, filename)
