+++ uploader/test/test_utils.py
@@ -1,4 +1,5 @@
 import os
+import re
 import tempfile
 from typing import Dict, List
 
@@ -17,6 +18,35 @@ from person.models import CofkUnionPerson
 from uploader.constants import MANDATORY_SHEETS
 from uploader.models import CofkCollectStatus, CofkCollectUpload
 
+spreadsheet_data = {'Work': [
+    [1, "test", "J", 1660, 1, 1, 1660, 1, 2, 1, 1, 1, 1, "test", "newton", 15257, "test", 1, 1, "test", "Wren",
+     22859, "test", 1, 1, "test", "Burford", 400285, "test", 1, 1, "Carisbrooke", 782, "test", 1, 1, "test",
+     "test", "fra;eng", '', '', '', '', '', '', "test", "test", "test", "Baskerville", 885, "test",
+     "test", "EMLO", "http://emlo.bodleian.ox.ac.uk/", "Early Modern Letters Online test"]],
+    'People': [["Baskerville", 885],
+               ["newton", 15257],
+               ["Wren", 22859]],
+    'Places': [['Burford', 400285],
+               ['Carisbrooke', 782]],
+    'Manifestation': [[1, 1, "ALS", 1, "Bodleian", "test", "test", '', '', '', '', ''],
+                      [2, 1, '', '', '', '', '', "P", "test", "test", '', '']],
+    'Repositories': [['Bodleian', 1]]}
+
+class MockMessages(object):
+    """
+    Messages middleware needs to be mocked because test factory bypasses middleware.
+    """
+    def add(self, *args):
+        return None
+
+    def __iter__(self):
+        for each in self.__dict__.values():
+            yield each
+
+
+
+upload_status = re.compile(r'Status: (?P<status>[\w|\s]+?) \| Number of works uploaded: (?P<works>\d+?) \|' \
+                r' Accepted: (?P<accepted>\d+?) \| Rejected: (?P<rejected>\d+?)')
 
 class UploaderTestCase(TestCase):
     def create_excel_file(self, data: Dict[str, List[List]] = None) -> str:
